Moved DDB Item back in normal scope
