+++ internal/log_analysis/alerts_api/table/table.go
@@ -20,10 +20,10 @@ package table
  */
 
 import (
+	"time"
+
 	"github.com/aws/aws-sdk-go/service/dynamodb"
 	"github.com/aws/aws-sdk-go/service/dynamodb/dynamodbiface"
-
-	"github.com/panther-labs/panther/api/lambda/alerts/models"
 )
 
 const (
@@ -35,9 +35,9 @@ const (
 
 // API defines the interface for the alerts table which can be used for mocking.
 type API interface {
-	GetAlert(*string) (*models.AlertItem, error)
-	ListByRule(string, *string, *int) ([]*models.AlertItem, *string, error)
-	ListAll(*string, *int) ([]*models.AlertItem, *string, error)
+	GetAlert(*string) (*AlertItem, error)
+	ListByRule(string, *string, *int) ([]*AlertItem, *string, error)
+	ListAll(*string, *int) ([]*AlertItem, *string, error)
 }
 
 // AlertsTable encapsulates a connection to the Dynamo alerts table.
@@ -53,3 +53,14 @@ var _ API = (*AlertsTable)(nil)
 
 // DynamoItem is a type alias for the item format expected by the Dynamo SDK.
 type DynamoItem = map[string]*dynamodb.AttributeValue
+
+// AlertItem is a DDB representation of an Alert
+type AlertItem struct {
+	AlertID      string    `json:"id"`
+	RuleID       string    `json:"ruleId"`
+	CreationTime time.Time `json:"creationTime"`
+	UpdateTime   time.Time `json:"updateTime"`
+	Severity     string    `json:"severity"`
+	EventCount   int       `json:"eventCount"`
+	LogTypes     []string  `json:"logTypes"`
+}
